+++ test/client/client_test.js
@@ -1,13 +1,14 @@
 'use strict';
 
 var chai = require('chai');
-
 var expect = chai.expect;
+
+var truthy = require('../../routes/ip_user.js');
 
 describe('client side test', function() {
   //its lame but it works
   it('should be true', function(){
-    expect(true).to.eql.true;
+    expect(truthy()).to.eql.true;
   });
 
 })
